ua,uag: split ua.c and uag.c
- uag.c contains code for User-Agent Group. E.g. functions that operate on the
  list of User-Agents.
- ua.c contains code for User-Agent. E.g. functions that operate on one User-
  Agent.
